refactor(acp): delete old custom ACP server, make SDK the only path
Remove the custom JSON-RPC ACP server (`acp_server.rs`) and the
`acp-sdk` feature flag. The SDK-based implementation is now always
compiled and is the sole ACP server path.

- Move `AcpError` into `acp_sdk_server.rs`
- Delete `acp_server.rs` (738 lines) and its unit tests
- Remove `[features]` sections from runtime and CLI Cargo.toml
- Make `agent-client-protocol` SDK deps non-optional
- Remove `#[cfg(feature = "acp-sdk")]` guards from main.rs
- Delete old `run_acp_server()`, `impl AcpAgent`, and duplicate
  slash command handler
- Remove integration test for old framed JSON-RPC protocol
