+++ rust/crates/runtime/Cargo.toml
@@ -5,17 +5,10 @@ edition.workspace = true
 license.workspace = true
 publish.workspace = true
 
-[features]
-acp-sdk = [
-    "dep:agent-client-protocol",
-    "dep:agent-client-protocol-schema",
-    "dep:agent-client-protocol-tokio",
-]
-
 [dependencies]
-agent-client-protocol = { workspace = true, optional = true }
-agent-client-protocol-schema = { workspace = true, optional = true }
-agent-client-protocol-tokio = { workspace = true, optional = true }
+agent-client-protocol = { workspace = true }
+agent-client-protocol-schema = { workspace = true }
+agent-client-protocol-tokio = { workspace = true }
 sha2 = "0.10"
 glob = "0.3"
 keyring = "3"
