+++ rust/crates/runtime/src/lib.rs
@@ -4,9 +4,7 @@
 //! MCP plumbing, tool-facing file operations, and the core conversation loop
 //! that drives interactive and one-shot turns.
 
-#[cfg(feature = "acp-sdk")]
 pub mod acp_sdk_server;
-pub mod acp_server;
 mod bash;
 pub mod bash_validation;
 mod bootstrap;
@@ -53,10 +51,7 @@ mod trust_resolver;
 mod usage;
 pub mod worker_boot;
 
-pub use acp_server::{
-    run_acp_server_with_io, run_acp_stdio_server, AcpAgent, AcpError, AcpServerError,
-    AcpServerOptions, AcpSessionUpdateObserver,
-};
+pub use acp_sdk_server::AcpError;
 pub use bash::{execute_bash, BashCommandInput, BashCommandOutput};
 pub use bootstrap::{BootstrapPhase, BootstrapPlan};
 pub use branch_lock::{detect_branch_lock_collisions, BranchLockCollision, BranchLockIntent};
