+++ rust/crates/rusty-sudocode-cli/Cargo.toml
@@ -5,9 +5,6 @@ edition.workspace = true
 license.workspace = true
 publish.workspace = true
 
-[features]
-acp-sdk = ["runtime/acp-sdk"]
-
 [[bin]]
 name = "scode"
 path = "src/main.rs"
